Amplifier getBuyingPrice test passed

+++ behaviours/Sellable.java
@@ -1,5 +1,5 @@
 package behaviours;
 
 public interface Sellable {
-  double calculateMarkup(int buyingPrice, int sellingPrice);
+  int calculateMarkup(int buyingPrice, int sellingPrice);
 }
